requirements.txt
aniso8601==8.0.0
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
flask-restplus==0.13.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
pycparser==2.19
pyOpenSSL==19.0.0
pyrsistent==0.15.4
pytz==2019.2
six==1.12.0
uuid==1.30
Werkzeug==0.15.6
